Moved UpsertPipe.apply -> SinkBuilder.upsertPipe and made it private
